+++ tests/testthat/test_add_absolute_targets.R
@@ -5,7 +5,7 @@ test_that("numeric(1)", {
   data(sim_projects, sim_actions, sim_features)
   # create problem
   p <- problem(sim_projects, sim_actions, sim_features,
-               "name", "success", "name", "cost", "name") %>%
+               "name", "success", "name", "cost", "name", FALSE) %>%
        add_absolute_targets(0.5)
   # calculate absolute targets
   targets <- p$targets$output()
@@ -25,7 +25,7 @@ test_that("numeric(4)", {
   data(sim_projects, sim_actions, sim_features)
   # create problem
   p <- problem(sim_projects, sim_actions, sim_features,
-               "name", "success", "name", "cost", "name") %>%
+               "name", "success", "name", "cost", "name", FALSE) %>%
        add_absolute_targets(seq_len(5) * 0.1)
   # calculate absolute targets
   targets <- p$targets$output()
@@ -46,7 +46,7 @@ test_that("character(1)", {
   sim_features$target <- seq_len(5) * 0.1
   # create problem
   p <- problem(sim_projects, sim_actions, sim_features,
-               "name", "success", "name", "cost", "name") %>%
+               "name", "success", "name", "cost", "name", FALSE) %>%
        add_absolute_targets("target")
   # calculate absolute targets
   targets <- p$targets$output()
@@ -64,7 +64,7 @@ test_that("character(1)", {
 test_that("invalid arguments", {
   data(sim_projects, sim_actions, sim_features)
   p <- problem(sim_projects, sim_actions, sim_features,
-               "name", "success", "name", "cost", "name")
+               "name", "success", "name", "cost", "name", FALSE)
   ## single numeric values
   expect_error({
     add_absolute_targets(p, 2)